+++ app/src/androidTest/java/edu/hm/cs/ma/todoguru/Util.kt
@@ -1,0 +1,31 @@
+package edu.hm.cs.ma.todoguru
+
+import android.view.View
+import android.view.ViewGroup
+import org.hamcrest.Description
+import org.hamcrest.Matcher
+import org.hamcrest.TypeSafeMatcher
+import java.time.LocalDate
+import java.time.format.DateTimeFormatter
+
+fun childAtPosition(
+    parentMatcher: Matcher<View>,
+    position: Int
+): Matcher<View> {
+
+    return object : TypeSafeMatcher<View>() {
+        override fun describeTo(description: Description) {
+            description.appendText("Child at position $position in parent ")
+            parentMatcher.describeTo(description)
+        }
+
+        public override fun matchesSafely(view: View): Boolean {
+            val parent = view.parent
+            return parent is ViewGroup && parentMatcher.matches(parent) && view == parent.getChildAt(
+                position
+            )
+        }
+    }
+}
+
+fun date(date: String) = LocalDate.parse(date, DateTimeFormatter.ofPattern("dd/MM/yyyy"))
